Refactored collision code on the player and added health system
Changed collision code to use b2ContactEvents insead of getting the b2ContactData
This is more robust, as when I tried to reuse the isGrounded code for an enemy hit
test, it returned garbage data for contactData, which was then throwing an error
when I called b2Shape_GetBody(id) as the id was garbage. To fix this, I get the
contacts from the world using b2World_GetContactEvents(pWorld). Then I can use the
resulting b2ContactEvents struct to get my contacts. I believe it is possible for
the player to be either shapeA or shapeB however, so I added code to make sure if
the player is one of the shapes, it is always shapeA for ease of coding. Then I
changed the isGrounded code slightly to work for multiple uses (hit by ground,
enemy, etc.). The result of this is a new TestContacts() function that replaces
isGround() and also the PlayerUpdate function that was created during the commit.

-Changed isGround to boolean
-Added TestContacts() func
-Added Hit() and Die() functions
-Added basic health system
-Changed PhysicsUserData to hold shouldDamage and damage for health system
-Updated sprite Atlas(.xcf, .png) with new ground sprite
-Changed player->UpdateBody() to be in loop()