Created camera class and platform rendering
-Added Camera class to types.h and replaced all calls to projViewMat with mainCamera->GetProjView()

-Added platform rendering:
-Added platform.vert and platform.frag to multiply the uv by the new instance attribute scaleFactor
-Added platformShader and platformAtlas sprite
-Platforms to be rendered will now be stored in the platforms vector instead of the sprites vector
-This, combined with the texture parameter GL_WRAP_S set to GL_REPEAT makes the new platform sprite
tile accross the length of the platform. (this probably won't work for vertical tiling).
NOTE: I should be able to just mod() the uv by 1.0 after doing this, but it produces an artefact
at the borders of the sprite on GTX 1080ti Driver: 560.94 instead if I just use texture wrapping,
and a platform atlas that is only one sprite wide, I can avoid this artefacting.
-Added another draw call to draw() to draw the platforms with their shader and instance attributes